chore: create BaseAirtableCommand
This change creates an abstract BaseAirtableCommand to have common
operations related to Airtable management commands.
